Add check for blank name coming from WP->Civi
We had an issue where names in Civi were getting overwritten by blank values. This happened if:

--the WP User had a blank name (somehow)
--the Civi user did not
--a password reset was requested

The reset triggers something which fires a profile sync, overwriting the name in Civi. 

I can't *think* of a situation where you'd want this to happen, so figured updating this plugin was the best approach. But happy to be corrected if you think it's not logical.